Refactor TextEditDialog into a separate file
